BugzID: 2666 - Implements Rad Social popup on event page
On the event detail page, there is now a share this button at the bottom of the description and when clicked, it pops up a social share modal.

I included the base social media icons (minus g+)

I am waiting on the proper copy of what the social icons should say, but the basic functionality is there.