BugzID: 3316 - Fixed the modal 404 and fixed some bugs within
Removed the modal js call from the view

Updated the view to put the class on the list so that it finds the next modal window. This will now open modals independently of each other.

Updated the js to grab the next modal window instead of just the modal class.