BugzID: 3018 - TRAF Homepage styling
Cleaned up some of the homepage styling for mobile views. The homepage was huge on mobile so I shortened the background image and made some of the text smaller so it all fits neatly on the homepage.

I also put a margin between the social and #traf on the insta feed.