BugzID: 2644 - Homepage announcement section styling fixed
The homepage announcement section looked nice but wasn't responsive enough to the content inside. I refactored it to look like the mockup and contain the content inside.

When it doesn't have content, the wave exists, but it is much thinner, and when it has content, it expands to contain that content with a reasonable amout of padding and margins to hold it.